move DataModelController.tsx in domain datamodel
